[breaking] Remove deprecated stuff (#117)
* [#113] Remove deprecated stuff

* [#113] Remove more deprecated stuff

* [#113] Move EbayFloatingLabel stories to EbayTextbox

* [#113] Update snapshots

* [#113] Move floating label hooks to common utils
