--- src/common/tooltip-utils/constants-legacy.ts
@@ -1,16 +1,0 @@
-import { PointerDirection, TooltipType } from './types'
-
-/**
- * @deprecated Use TooltipType union type instead
- */
-export const TOOLTIP_TYPES: TooltipType[] = [`tooltip`, `infotip`, `tourtip`]
-
-/**
- * @deprecated Use PointerDirection union type instead
- */
-export const POINTER_TYPES: PointerDirection[] = [
-    `top`, `top-left`, `top-right`,
-    `right`, `right-bottom`, `right-top`,
-    `bottom`, `bottom-left`, `bottom-right`,
-    `left`, `left-bottom`, `left-top`
-]
