+++ src/common/tooltip-utils/index.ts
@@ -4,4 +4,3 @@ export { default as TooltipContent } from './tooltip-content'
 export { useTooltip } from './use-tooltip'
 export { TooltipType, PointerDirection } from './types'
 export { DEFAULT_POINTER_DIRECTION, POINTER_STYLES } from './constants'
-export { POINTER_TYPES, TOOLTIP_TYPES } from './constants-legacy'
