--- src/ebay-dialog/README.md
@@ -1,4 +1,0 @@
-# EbayDialog
-
-## Deprecated
-Use [EbayFullscreenDialog](../ebay-fullscreen-dialog), [EbayLightboxDialog](../ebay-lightbox-dialog), [EbayPanelDialog](../ebay-panel-dialog), [EbayDrawerDialog](../ebay-drawer-dialog) instead.
