+++ src/ebay-field/index.ts
@@ -2,4 +2,3 @@ export { default as EbayField } from './field'
 export { default as EbayLabel, Props as EbayLabelProps } from './label'
 export { default as EbayFieldDescription } from './description'
 export { FieldDescriptionType, FieldDescriptionPosition, FieldLayoutType, LabelPosition } from './types'
-export { fieldLayoutType, fieldDescriptionType, fieldDescriptionPosition } from './enums'
