--- src/ebay-floating-label/index.ts
@@ -1,2 +1,0 @@
-export { default as EbayFloatingLabel } from './floating-label'
-export { Size } from '../ebay-textbox'
