--- src/ebay-infotip/enums.ts
@@ -1,9 +1,0 @@
-import { Variant } from './types'
-
-/**
- * @deprecated Use Variant union type instead
- */
-export const variants: { [key: string]: Variant } = {
-    DEFAULT: `default`,
-    MODAL: `modal`
-}
