+++ src/ebay-infotip/index.ts
@@ -3,4 +3,3 @@ export { default as EbayInfotipHost } from './ebay-infotip-host'
 export { default as EbayInfotipHeading } from './ebay-infotip-heading'
 export { default as EbayInfotipContent } from './ebay-infotip-content'
 export { Variant } from './types'
-export { variants } from './enums'
