--- src/ebay-legacy-floating-label/index.ts
@@ -1,1 +1,0 @@
-export { default as EbayLegacyFloatingLabel } from './legacy-floating-label'
