--- src/ebay-notice-content/README.md
@@ -1,6 +1,0 @@
-# DEPRECATED
-Use those instead:
-- `import { EbayNoticeContent } from '@ebay/ui-core-react/ebay-notice-base`
-- `import { EbayNoticeContent } from '@ebay/ui-core-react/ebay-page-notice`
-- `import { EbayNoticeContent } from '@ebay/ui-core-react/ebay-section-notice`
-- `import { EbayNoticeContent } from '@ebay/ui-core-react/ebay-inline-notice`
