--- src/ebay-notice-content/index.ts
@@ -1,1 +1,0 @@
-export { EbayNoticeContent } from '../ebay-notice-base'
