--- src/ebay-tab/README.md
@@ -1,1 +1,0 @@
-# Deprecated, use [EbayTabs](../ebay-tabs) instead.
