--- src/ebay-tab/index.ts
@@ -1,5 +1,0 @@
-/**
- * @deprecated Use EbayTabs instead
- */
-export { EbayTabs as EbayTab, EbayTab as EbayTabHeading, EbayTabPanel } from '../ebay-tabs'
-export { activations, sizes } from './enums'
