+++ src/ebay-textbox/textbox.tsx
@@ -2,7 +2,7 @@ import React, { ChangeEvent, cloneElement, ComponentProps, FC, FocusEvent, Ref, 
 import classNames from 'classnames'
 import { findComponent, withForwardRef } from '../common/component-utils'
 import { EbayTextboxPostfixIcon, EbayTextboxPrefixIcon, Size } from './index'
-import { useFloatingLabel } from '../ebay-floating-label/hooks'
+import { useFloatingLabel } from '../common/floating-label-utils/hooks'
 
 const isControlled = value => typeof value !== 'undefined'
 
@@ -12,7 +12,6 @@ export type EbayTextboxProps = {
     fluid?: boolean;
     invalid?: boolean;
     multiline?: boolean;
-    underline?: boolean;
     defaultValue?: string;
     inputSize?: Size;
     onFocus?: (e?: FocusEvent<HTMLTextAreaElement & HTMLInputElement>, value?: string) => void;
@@ -28,7 +27,6 @@ const EbayTextbox: FC<EbayTextboxProps> = ({
     invalid,
     fluid,
     multiline,
-    underline /* DEPRECATED */,
     onChange = () => {},
     onFocus = () => {},
     onBlur = () => {},
@@ -96,7 +94,6 @@ const EbayTextbox: FC<EbayTextboxProps> = ({
 
     const inputClassName = classNames('textbox__control', {
         'textbox__control--fluid': fluid,
-        'legacy-textbox-underline': underline,
         'textbox__control--large': inputSize === 'large'
     })
     const wrapperClassName = classNames('textbox', {
