+++ src/ebay-tooltip/index.ts
@@ -2,4 +2,3 @@ export { default as EbayTooltip } from './ebay-tooltip'
 export { default as EbayTooltipContent } from './ebay-tooltip-content'
 export { default as EbayTooltipHost } from './ebay-tooltip-host'
 export { PointerDirection, Tooltip } from '../common/tooltip-utils'
-export { POINTER_TYPES, TOOLTIP_TYPES } from '../common/tooltip-utils'
