+++ src/index.ts
@@ -31,17 +31,11 @@ export { EbayLightboxDialog } from './ebay-lightbox-dialog'
 export { EbayPanelDialog } from './ebay-panel-dialog'
 export { EbayTextbox, EbayTextboxPrefixIcon, EbayTextboxPostfixIcon, EbayTextboxProps } from './ebay-textbox'
 export { EbaySelect, EbaySelectOption } from './ebay-select'
-export { EbayFloatingLabel } from './ebay-floating-label'
-export {
-    EbayTooltip, EbayTooltipContent, EbayTooltipHost, PointerDirection, Tooltip,
-    POINTER_TYPES, TOOLTIP_TYPES
-} from './ebay-tooltip'
-export { EbayInfotip, EbayInfotipHost, EbayInfotipContent, EbayInfotipHeading, Variant, variants } from './ebay-infotip'
-export { EbayLegacyFloatingLabel } from './ebay-legacy-floating-label'
+export { EbayTooltip, EbayTooltipContent, EbayTooltipHost, PointerDirection, Tooltip } from './ebay-tooltip'
+export { EbayInfotip, EbayInfotipHost, EbayInfotipContent, EbayInfotipHeading, Variant } from './ebay-infotip'
 export {
     EbayField, EbayLabel, EbayFieldDescription,
-    FieldLayoutType, FieldDescriptionType, FieldDescriptionPosition, LabelPosition,
-    fieldLayoutType, fieldDescriptionType, fieldDescriptionPosition
+    FieldLayoutType, FieldDescriptionType, FieldDescriptionPosition, LabelPosition
 } from './ebay-field'
 export {
     EbayProgressStepper, EbayProgressStep, EbayProgressTitle, StepperDirection, StepState
